Added remove books from local storage
