+++ modules/setLocalStorage.js
@@ -1,0 +1,7 @@
+import { getLocalStorageBooks } from "./getLocalStorage.js";
+export function setLocalStorage(book) {
+    const books = getLocalStorageBooks();
+    books.push(book);
+    localStorage.setItem('books', JSON.stringify(books));
+
+  }
